qnap_boot

+++ luci-app-cowbping/root/etc/cowbping.sh
@@ -1,38 +1,22 @@
 #!/bin/sh
 # Author: wulishui <wulishui@gmail.com>
 
-echo "-----Author: wulishui , 2020627 v1.7-----" > /tmp/log/cowbping.log
-time=$(uci get cowbping.cowbping.time 2>/dev/null)
-delaytime=$(uci get cowbping.cowbping.delaytime 2>/dev/null)
-sleep "$delaytime"
+log_limit() {
+logsize=`du /tmp/log/cowbping.log 2>/dev/null|awk '{print $1}'`
+[ $logsize -gt 80 ] || return
+cat /tmp/log/cowbping.log >> /tmp/log/cowbping.log_
+echo ">"$(date +"%Y-%m-%d %H:%M:%S")" ：日志文件过大，旧的记录已暂时经转移到/tmp/log/cowbping.log_。" > /tmp/log/cowbping.log 2>/dev/null
+}
 
-while :
-do
-work_mode=$(uci get cowbping.cowbping.work_mode 2>/dev/null)
-sum=$(uci get cowbping.cowbping.sum 2>/dev/null)
-address=$(uci get cowbping.cowbping.address 2>/dev/null)
-pkglost=$(uci get cowbping.cowbping.pkglost 2>/dev/null)
-
-ping=`ping -c 1 "$address"|grep -o -E "([0-9]|[1-9][0-9]|100)"% | awk -F '%' '{print $1}'`
-if [ "$ping" -ge "$pkglost" ]; then
- sum0=1
- for i in $(seq 1 $((sum-1)))
- do
-   ping=`ping -c 1 "$address"|grep -o -E "([0-9]|[1-9][0-9]|100)"% | awk -F '%' '{print $1}'`
-   [ "$ping" -ge "$pkglost" ] && sum0=$((sum0+1))
- done
-else
- sum0=0
-fi
-
-if [ "$sum0" -ge "$sum" ]; then
-echo "20"$(date +"%y-%m-%d %H:%M:%S")"----断网啦!!!" >> /tmp/log/cowbping.log
+CMDS() {
+log_limit
+echo ""$(date +"%Y-%m-%d %H:%M:%S")"----断网啦!!!" >> /tmp/log/cowbping.log
 case "$work_mode" in
 "1")
 reboot
 ;;
 "2")
-killall -q pppd && pppd file /tmp/ppp/options.wan0 2>/dev/null
+killall -q pppd && sleep 5 && pppd file /tmp/ppp/options.wan 2>/dev/null
 ;;
 "3")
 wifi down && wifi up 2>/dev/null
@@ -46,19 +30,38 @@ eval ${command} 2>/dev/null
 ;;
 "6")
 MAC=`dd if=/dev/urandom bs=1 count=32 2>/dev/null | md5sum | cut -b 0-12 | sed 's/\(..\)/\1:/g; s/.$//'`
+wifi down
 uci set wireless.@wifi-iface[0].macaddr="$MAC"
 uci commit wireless
-wifi down
 wifi up
 ;;
 "7")
 poweroff 2>/dev/null
 ;;
 esac
-fi
+}
 
+echo "-----Author: wulishui , 20201003 v2.0-----" >> /tmp/log/cowbping.log
+delaytime=$(uci get cowbping.cowbping.delaytime 2>/dev/null)
+sleep "$delaytime"
+time=$(uci get cowbping.cowbping.time 2>/dev/null)
+work_mode=$(uci get cowbping.cowbping.work_mode 2>/dev/null)
+sum=$(uci get cowbping.cowbping.sum 2>/dev/null)
+address=$(uci get cowbping.cowbping.address 2>/dev/null)
+pkglost=$(uci get cowbping.cowbping.pkglost 2>/dev/null)
+while :
+do
+ping=`ping -c 1 "$address"|grep -o -E "([0-9]|[1-9][0-9]|100)"% | awk -F '%' '{print $1}'`
+if [ "$ping" -ge "$pkglost" ]; then
+ sum0=1
+ for i in $(seq 1 $((sum-1)))
+ do
+   ping=`ping -c 1 "$address"|grep -o -E "([0-9]|[1-9][0-9]|100)"% | awk -F '%' '{print $1}'` ; [ "$ping" -ge "$pkglost" ] && sum0=$((sum0+1))
+  [ "$sum0" -ge "$sum" ] && CMDS
+ done
+ [ "$sum" == 1 ] && CMDS
+fi
 sleep "$time"
 done
-
